csf_server/requirements.txt
Django==2.1.7
django-rest-framework==0.1.0
djangorestframework==3.9.1
pytz==2018.9
django-cors-headers==3.0.2
gunicorn==19.9.0